+++ source/src/CharacterConfigVM.cs
@@ -9,7 +9,7 @@ namespace EnhancedBattleTest
 {
     public abstract class CharacterConfigVM : ViewModel
     {
-        public abstract void SetConfig(CharacterConfig config, bool isAttacker);
+        public abstract void SetConfig(TeamConfig teamConfig, CharacterConfig config, bool isAttacker);
         public abstract void SelectedCharacterChanged(Character character);
 
         public static CharacterConfigVM Create(bool isMultiplayer)