+++ source/src/MPCharacterConfigVM.cs
@@ -26,6 +26,8 @@ namespace EnhancedBattleTest
 
 
         public TextVM IsHeroText { get; }
+
+        public TextVM MaleRatioText { get; }
         public TextVM FemaleRatioText { get; }
         public BoolVM IsHero { get; }
         public NumberVM<float> FemaleRatio { get; }
@@ -36,6 +38,7 @@ namespace EnhancedBattleTest
             SecondPerks = new SelectorVM<SelectorItemVM>(0, null);
 
             IsHeroText = new TextVM(GameTexts.FindText("str_ebt_is_hero"));
+            FemaleRatioText = new TextVM(GameTexts.FindText("str_ebt_male_ratio"));
             FemaleRatioText = new TextVM(GameTexts.FindText("str_ebt_female_ratio"));
             IsHero = new BoolVM(_config.IsHero);
             FemaleRatio = new NumberVM<float>(_config.FemaleRatio, 0, 1, false);
@@ -51,7 +54,7 @@ namespace EnhancedBattleTest
             };
         }
 
-        public override void SetConfig(CharacterConfig config, bool isAttacker)
+        public override void SetConfig(TeamConfig teamConfig, CharacterConfig config, bool isAttacker)
         {
             if (!(config is MPCharacterConfig mpConfig))
                 return;