+++ source/src/TroopGroup.cs
@@ -26,14 +26,14 @@ namespace EnhancedBattleTest
             }
         }
 
-        public TroopGroup(TroopGroupConfig config, bool isPlayerSide, BattleTypeConfig battleTypeConfig)
+        public TroopGroup(TeamConfig teamConfig, TroopGroupConfig config, bool isPlayerSide, BattleTypeConfig battleTypeConfig)
         {
             _config = config;
             Troops = new MBBindingList<TroopVM>();
             for (int i = 0; i < _config.Troops.Length; ++i)
             {
                 GameTexts.SetVariable("TroopIndex", i);
-                Troops.Add(new TroopVM(_config.Troops[i], GameTexts.FindText("str_ebt_troop_role", "Soldiers"),
+                Troops.Add(new TroopVM(teamConfig, _config.Troops[i], GameTexts.FindText("str_ebt_troop_role", "Soldiers"),
                     isPlayerSide, battleTypeConfig));
             }
         }